+++ indico/modules/events/reminders/util.py
@@ -20,7 +20,7 @@ from indico.modules.events.models.events import EventType
 from indico.web.flask.templating import get_template_module
 
 
-def make_reminder_email(event, with_agenda, note):
+def make_reminder_email(event, with_agenda, with_description, note):
     """Returns the template module for the reminder email.
 
     :param event: The event
@@ -31,4 +31,5 @@ def make_reminder_email(event, with_agenda, note):
         with_agenda = False
     agenda = event.timetable_entries.filter_by(parent_id=None).all() if with_agenda else None
     return get_template_module('events/reminders/emails/event_reminder.txt', event=event,
-                               url=event.short_external_url, note=note, with_agenda=with_agenda, agenda=agenda)
+                               url=event.short_external_url, note=note, with_agenda=with_agenda,
+                               with_description=with_description, agenda=agenda)